Reorganize response module
